Clean unused code (#2174)
* Remove unused vercel code and cookie parser

* Remove app-init from Job Launcher and Reputation Oracle
