+++ packages/apps/fortune/exchange-oracle/server/package.json
@@ -31,7 +31,6 @@
     "@nestjs/axios": "^3.0.2",
     "@nestjs/common": "^10.2.7",
     "@nestjs/core": "^10.2.8",
-    "@nestjs/platform-express": "^10.3.9",
     "joi": "^17.12.2",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.2.0"
