+++ packages/apps/fortune/recording-oracle/package.json
@@ -28,11 +28,9 @@
     "@nestjs/common": "^10.2.7",
     "@nestjs/config": "^3.1.1",
     "@nestjs/core": "^10.2.8",
-    "@nestjs/platform-express": "^10.3.9",
     "@nestjs/swagger": "^7.1.13",
     "body-parser": "^1.20.2",
     "class-validator": "^0.14.1",
-    "cookie-parser": "^1.4.6",
     "helmet": "^7.1.0",
     "reflect-metadata": "^0.1.13"
   },
