+++ packages/apps/job-launcher/server/.env.example
@@ -3,7 +3,6 @@ NODE_ENV=development
 HOST=localhost
 PORT=3000
 FE_URL=http://localhost:3001
-SESSION_SECRET=test
 
 # Database
 POSTGRES_HOST=0.0.0.0
