+++ packages/apps/job-launcher/server/package.json
@@ -35,21 +35,16 @@
     "@nestjs/core": "^10.2.8",
     "@nestjs/jwt": "^10.2.0",
     "@nestjs/passport": "^10.0.0",
-    "@nestjs/platform-express": "^10.3.9",
     "@nestjs/schedule": "^4.0.1",
     "@nestjs/serve-static": "^4.0.1",
     "@nestjs/swagger": "^7.1.13",
     "@nestjs/terminus": "^10.2.1",
     "@nestjs/typeorm": "^10.0.1",
     "@sendgrid/mail": "^8.1.3",
-    "@types/cookie-parser": "^1.4.3",
-    "@types/express-session": "^1.17.10",
     "@types/passport-jwt": "^4.0.1",
     "bcrypt": "^5.1.1",
     "class-transformer": "^0.5.1",
-    "cookie-parser": "^1.4.6",
     "decimal.js": "^10.4.3",
-    "express-session": "^1.17.3",
     "joi": "^17.12.2",
     "json-stable-stringify": "^1.1.1",
     "nestjs-minio-client": "^2.2.0",
