--- packages/apps/job-launcher/server/vercel.json
@@ -1,97 +1,0 @@
-{
-  "version": 2,
-  "buildCommand": "yarn workspace @human-protocol/sdk build && yarn build",
-  "outputDirectory": "dist",
-  "functions": {
-    "api/app.ts":{
-      "maxDuration": 300
-    }
-  },
-  "redirects": [
-    { "source": "/", "destination": "/swagger" }
-  ],
-  "rewrites": [
-    {
-      "source": "/swagger",
-      "destination": "/api/app.ts"
-    },
-    {
-      "source": "/swagger/(.*)",
-      "destination": "/api/app.ts"
-    },
-    {
-      "source": "/health",
-      "destination": "/api/app.ts"
-    },
-    {
-      "source": "/auth/(.*)",
-      "destination": "/api/app.ts"
-    },
-    {
-      "source": "/user/balance",
-      "destination": "/api/app.ts"
-    },
-    {
-      "source": "/payment/(.*)",
-      "destination": "/api/app.ts"
-    },
-    {
-      "source": "/web3/(.*)",
-      "destination": "/api/app.ts"
-    },
-    {
-      "source": "/job/(.*)",
-      "destination": "/api/app.ts"
-    },
-    {
-      "source": "/webhook",
-      "destination": "/api/app.ts"
-    },
-    {
-      "source": "/cron/(.*)",
-      "destination": "/api/app.ts"
-    }
-  ],
-  "crons": [
-    {
-      "path": "/cron/escrow/create",
-      "schedule": "*/2 * * * *"
-    },
-    {
-      "path": "/cron/escrow/setup",
-      "schedule": "1-59/2 * * * *"
-    },
-    {
-      "path": "/cron/escrow/fund",
-      "schedule": "*/2 * * * *"
-    },
-    {
-      "path": "/cron/escrow/cancel",
-      "schedule": "*/5 * * * *"
-    },
-    {
-      "path": "/cron/webhook/process",
-      "schedule": "*/3 * * * *"
-    }
-  ],
-  "headers": [
-    {
-      "source": "/(.*)",
-      "headers": [
-        {
-          "key": "Access-Control-Allow-Origin",
-          "value": "*"
-        },
-        {
-          "key": "Access-Control-Allow-Methods",
-          "value": "*"
-        },
-        {
-          "key": "Access-Control-Allow-Headers",
-          "value": "*"
-        }
-      ]
-    }
-  ],
-  "ignoreCommand": "git diff HEAD^ HEAD --quiet ."
-}
