+++ packages/apps/reputation-oracle/server/package.json
@@ -37,7 +37,6 @@
     "@nestjs/core": "^10.2.8",
     "@nestjs/jwt": "^10.2.0",
     "@nestjs/passport": "^10.0.0",
-    "@nestjs/platform-express": "^10.3.9",
     "@nestjs/schedule": "^4.0.1",
     "@nestjs/swagger": "^7.1.13",
     "@nestjs/terminus": "^10.2.1",
