--- packages/apps/reputation-oracle/server/src/app-init.ts
@@ -1,23 +1,0 @@
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
-import { json, urlencoded } from 'body-parser';
-import { useContainer } from 'class-validator';
-import helmet from 'helmet';
-
-import { AppModule } from './app.module';
-
-export default async function init(app: any) {
-  useContainer(app.select(AppModule), { fallbackOnErrors: true });
-  app.use(json({ limit: '5mb' }));
-  app.use(urlencoded({ limit: '5mb', extended: true }));
-
-  const config = new DocumentBuilder()
-    .addBearerAuth()
-    .setTitle('Reputation Oracle API')
-    .setDescription('Swagger Reputation Oracle API')
-    .setVersion('1.0')
-    .build();
-  const document = SwaggerModule.createDocument(app, config);
-  SwaggerModule.setup('swagger', app, document);
-
-  app.use(helmet());
-}
