+++ packages/apps/reputation-oracle/server/src/main.ts
@@ -1,16 +1,32 @@
-import { NestFactory } from '@nestjs/core';
-import { AppModule } from './app.module';
 import { INestApplication } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
+import { NestFactory } from '@nestjs/core';
+import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
+import { json, urlencoded } from 'body-parser';
+import helmet from 'helmet';
+import { AppModule } from './app.module';
+import { useContainer } from 'class-validator';
 import { ServerConfigService } from './common/config/server-config.service';
-import init from './app-init';
 
 async function bootstrap() {
   const app = await NestFactory.create<INestApplication>(AppModule, {
     cors: true,
   });
+  useContainer(app.select(AppModule), { fallbackOnErrors: true });
 
-  await init(app);
+  app.use(json({ limit: '5mb' }));
+  app.use(urlencoded({ limit: '5mb', extended: true }));
+
+  const config = new DocumentBuilder()
+    .addBearerAuth()
+    .setTitle('Reputation Oracle API')
+    .setDescription('Swagger Reputation Oracle API')
+    .setVersion('1.0')
+    .build();
+  const document = SwaggerModule.createDocument(app, config);
+  SwaggerModule.setup('swagger', app, document);
+
+  app.use(helmet());
 
   const configService: ConfigService = app.get(ConfigService);
   const serverConfigService = new ServerConfigService(configService);
