--- packages/apps/reputation-oracle/server/vercel.json
@@ -1,85 +1,0 @@
-{
-  "version": 2,
-  "buildCommand": "yarn workspace @human-protocol/sdk build && yarn build",
-  "outputDirectory": "dist",
-  "functions": {
-    "api/app.ts":{
-      "maxDuration": 300
-    }
-  },
-  "redirects": [
-    { "source": "/", "destination": "/swagger" }
-  ],
-  "rewrites": [
-    {
-      "source": "/swagger",
-      "destination": "/api/app.ts"
-    },
-    {
-      "source": "/swagger/(.*)",
-      "destination": "/api/app.ts"
-    },
-    {
-      "source": "/health",
-      "destination": "/api/app.ts"
-    },
-    {
-      "source": "/reputation(.*)",
-      "destination": "/api/app.ts"
-    },
-    {
-      "source": "/web3/prepare-signature",
-      "destination": "/api/app.ts"
-    },
-    {
-      "source": "/webhook",
-      "destination": "/api/app.ts"
-    },
-    {
-      "source": "/auth/(.*)",
-      "destination": "/api/app.ts"
-    },
-    {
-      "source": "/user/(.*)",
-      "destination": "/api/app.ts"
-    },
-    {
-      "source": "/kyc/(.*)",
-      "destination": "/api/app.ts"
-    },
-    {
-      "source": "/cron/(.*)",
-      "destination": "/api/app.ts"
-    }
-  ],
-  "crons": [
-    {
-      "path": "/cron/webhook/pending",
-      "schedule": "*/5 * * * *"
-    },
-    {
-      "path": "/cron/webhook/paid",
-      "schedule": "*/5 * * * *"
-    }
-  ],
-  "headers": [
-    {
-      "source": "/(.*)",
-      "headers": [
-        {
-          "key": "Access-Control-Allow-Origin",
-          "value": "*"
-        },
-        {
-          "key": "Access-Control-Allow-Methods",
-          "value": "*"
-        },
-        {
-          "key": "Access-Control-Allow-Headers",
-          "value": "*"
-        }
-      ]
-    }
-  ],
-  "ignoreCommand": "git diff HEAD^ HEAD --quiet ."
-}
